requirements.txt
python-dotenv~=1.1.0
SQLAlchemy~=2.0.40
passlib~=1.7.4
pip~=23.2.1
uvicorn[standard]
wheel~=0.41.2
setuptools~=68.2.0
greenlet~=3.2.0
typing_extensions~=4.13.2
pydantic~=2.11.3
fastapi~=0.115.12
dishka~=1.5.3
psycopg2-binary
asyncpg
alembic
python-multipart